optimize memory consumption
primarily avoids holding many copies of the entityIDs. they are fairly
long (42 characters typically, which is ~120bytes) and tend to be read
from everywhere, resulting in lots of different instances of identical
strings. using a canonical instance of each entityID String avoids
storing a separate copy for each nethash (in the ranking cache) and for
each entityID/nethash tuple (in the login logger).

this makes the cache sizes very predictable, and relatively small
because the caches only really hold references. thus the size limits
can be set quite generously, significantly reducing the impact of high
load situations.