validate pool connections before using
this is slightly less efficient, but it ensures that a new connection from
the pool will never be dead right away. the code still has to handle the
boundary case where the connection breaks during a request (and retry), but
it can at least be sure that retrying will actually work.