improve filter update delay handling
use a much longer delay when there is no fallback list of filters.
returning an older list is perfectly fine most of the time, but not
filtering the list at all is very bad for usability. also keep
filters around for much longer so that the shorter delay can be used
as frequently as possible.