XSLT-transform metadata before parsing
this sounds complicated and slower, but XPath queries on a DOM tree are
massively slower than XSLT transforms on an XML stream. also, the resulting
intermediate XML is so simple that it's more readable to just parse it
using getChildNodes() than with XPath queries.

real-life implication is that metadata parsing time goes down from 3~5
seconds to a few hundred milliseconds.