only use filter text if it isn't the dummy text
there are situations, such as pressing Ctrl-F in firefox, where the key*
handlers fire, but simultaneously the box is unfocused. thus it would be
filled with the dummy text, and because the filtering is triggered as well,
it would filter by the dummy text. because that doesn't match any IdP, the
user gets an annoying empty list.