trigger search on focus/blur as well
there are ways to change the text field that don't trigger any of the key*
handlers, such as "Cut" from the context menu, or a "helpful" close button
that IE renders in the text box; these will not update the search results.
bind on focus / blur so that at least (un-)focusing the box updates the
search results.

of course, none of this would be a problem if the "change" event would
work in real time...