fix boundary case leading to ignored favorites
if the first request after startup comes in before metadata is available,
IdPMetaParser would filter out everything, leaving an empty list of IP-hash
favorites. this then gets cached, leaving only the cookie favorite and the
global list for a while.

throwing an exception in that case leads to the same result (no IP-hash
favorites being displayed), but doesn't get cached. therefore, as soon as
metadata becomes available, the Ip-hash favorites will appear in the list.